chỉnh sửa lại payment
chỉnh sửa lại payment

+++ SWP291_Team2_SE1829NJ/src/java/model/BookingItem.java
@@ -67,7 +67,7 @@ public class BookingItem {
         this.formattedEndDate = formattedEndDate;
     }
 
-    public double calculateCost() {
+    public float calculateCost() {
         long days = java.time.temporal.ChronoUnit.DAYS.between(startDate, endDate);
         return days * room.getPrice() * quantity;
     }
